Remove unused/unnecessary util functions.
