--- pkg/utils/json.go
@@ -1,41 +1,0 @@
-// This file is part of MinIO Direct CSI
-// Copyright (c) 2021 MinIO, Inc.
-//
-// This program is free software: you can redistribute it and/or modify
-// it under the terms of the GNU Affero General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-//
-// This program is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU Affero General Public License for more details.
-//
-// You should have received a copy of the GNU Affero General Public License
-// along with this program.  If not, see <http://www.gnu.org/licenses/>.
-
-package utils
-
-import (
-	"fmt"
-
-	jsonFormatter "encoding/json"
-)
-
-// MustJSON converts given value to JSON string.
-func MustJSON(obj interface{}) string {
-	j, err := ToJSON(obj)
-	if err != nil {
-		panic(err)
-	}
-	return j
-}
-
-// ToJSON converts given value to JSON string.
-func ToJSON(obj interface{}) (string, error) {
-	formattedObj, err := jsonFormatter.MarshalIndent(obj, "", "  ")
-	if err != nil {
-		return "", fmt.Errorf("error marshaling to JSON: %v", err)
-	}
-	return string(formattedObj), nil
-}
