+++ pkg/utils/utils_test.go
@@ -107,56 +107,3 @@ func TestNewSafeFile(t *testing.T) {
 	}
 
 }
-
-func TestVolumeStatusTransitions(t1 *testing.T) {
-
-	statusList := []metav1.Condition{
-		{
-			Type:               "staged",
-			Status:             metav1.ConditionFalse,
-			LastTransitionTime: metav1.Date(2000, 1, 1, 0, 0, 0, 0, time.UTC),
-		},
-		{
-			Type:               "published",
-			Status:             metav1.ConditionFalse,
-			LastTransitionTime: metav1.Date(2000, 1, 2, 0, 0, 0, 0, time.UTC),
-		},
-	}
-
-	testCases := []struct {
-		name       string
-		condType   string
-		condStatus metav1.ConditionStatus
-	}{
-		{
-			name:       "NodeStageVolumeTransition",
-			condType:   "staged",
-			condStatus: metav1.ConditionTrue,
-		},
-		{
-			name:       "NodePublishVolumeTransition",
-			condType:   "published",
-			condStatus: metav1.ConditionTrue,
-		},
-		{
-			name:       "NodeUnpublishVolumeTransition",
-			condType:   "published",
-			condStatus: metav1.ConditionFalse,
-		},
-		{
-			name:       "NodeUnstageVolumeTransition",
-			condType:   "staged",
-			condStatus: metav1.ConditionFalse,
-		},
-	}
-
-	for _, tt := range testCases {
-		t1.Run(tt.name, func(t1 *testing.T) {
-			UpdateCondition(statusList, tt.condType, tt.condStatus, "", "")
-			if !IsCondition(statusList, tt.condType, tt.condStatus, "", "") {
-				t1.Errorf("Test case name %s: Status transition failed (Type, Status) = (%s, %v) condition list: %v", tt.name, tt.condType, tt.condStatus, statusList)
-			}
-		})
-	}
-
-}
